Added the correct return type (int) and function parameter to the attack function along with the statements for the function body. Will use a const for the function parameter as it will not change the value within the function itself
Added parameter (nonconst) to the defend function along with the statements for the function body. Will use a nonconst object for the function parameter as it will change the value within the function itself and an integer damage variable

Added parameter (nonconst) to the heal function along with the statements for the function body. Will use a nonconst object for the function parameter as it will change the value within the function itself