Menu driven bug fix:
The monsterAttack() command is the original function call and works fine. However, a bug is produced if the user does not enter an 'a' or 'h'. The bug happens when the user enters something other than 'a' or 'h' and subsequently the monsters attack despite the invalid menu selection. To solve this I added an if statement to prevent the bug. It could have also been implemented into the switch case.